# Conflicts:
#	lib/ui/calorie_tunes.dart
#	lib/ui/device_usages.dart
#	lib/ui/leaderboards/leaderboard_device_hub.dart
#	lib/ui/power_tunes.dart

+++ test/cadence_mixin_test.dart
@@ -20,7 +20,7 @@ void main() {
     final cadenceMixin = CadenceMixin();
 
     expect(cadenceMixin.cadenceData.isEmpty, true);
-    expect(cadenceMixin.computeCadence(), 0);
+    expect(cadenceMixin.computeCadence().toInt(), 0);
   });
 
   group('Cadence Mixin clearCadenceData clears cadence data', () {
@@ -37,7 +37,7 @@ void main() {
 
         cadenceMixin.clearCadenceData();
         expect(cadenceMixin.cadenceData.isEmpty, true);
-        expect(cadenceMixin.computeCadence(), 0);
+        expect(cadenceMixin.computeCadence().toInt(), 0);
       });
     }
   });
